Organize folders

+++ CUnderline/Homework/icoding/Experience1/CalculateTax.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    double a;
+    float a;
     scanf("%lf", &a);
     a = a * 1.05;
     printf("%lf\n", a);
