+++ CUnderline/Homework/icoding/Experience1/LoanBalance.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    double a, b, c, e, f, g, r;
+    float a, b, c, e, f, g, r;
     scanf("%lf", &a); // a是贷款金额
     scanf("%lf", &b); // b是利率
     scanf("%lf", &c); // c是每月还款金额
