+++ CUnderline/Learning/Download/231030函数/decompose.c
@@ -2,22 +2,22 @@
 
 #include <stdio.h>
 
-void decompose(double d, int *i, double *f) {
-    *i = (int)d; // 强制类型转换
+void decompose(float d, int *i, float *f) {
+    *i = (int) d; // 强制类型转换
     *f = d - *i;
 }
 
 int main() {
-    double d;
+    float d;
     scanf("%lf", &d);
 
     /*
     int *i;
-    double *f;
+    float *f;
     decompse(d, i, f);  // 错误：i和f都是野指针
     */
     int i;
-    double f;
+    float f;
     decompose(d, &i, &f);
     printf("d=%lf, i=%d, f=%lf\n", d, i, f);
