+++ CUnderline/Practice/decompose.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
-void decompose(double d, int* i, double* f) {
+void decompose(float d, int* i, float* f) {
     *i = (int) d;
     *f = d - *i;
 }
 int main() {
     int i;
-    double f, d;
+    float f, d;
     scanf("%lf", &d);
     decompose(d, &i, &f);
     printf("%d,%.3f\n", i, f);
